+++ packages/hydrogen/src/hooks/useNavigate/useNavigate.ts
@@ -1,4 +1,4 @@
-import {useRouter} from '../../foundation/Router/Router';
+import {useRouter} from '../../foundation/Router/BrowserRouter.client';
 
 type NavigationOptions = {
   /** Whether to update the state object or URL of the current history entry. Default to false */
